fix(services): :bug: FIX createItemService
The issue with your createItemService function is in the structure of the axios.post request. Axios expects the data to be passed as the second argument and the configuration (like headers) to be passed as the third argument. In your code, you’re passing both data and headers in a single object, which isn’t correct.